Fixed some issues
